AWS refactor
