+++ quest_test/test_persistence.py
@@ -1,7 +1,6 @@
 import asyncio
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from typing import Callable
 from dotenv import load_dotenv
 
 import pytest
@@ -11,8 +10,17 @@ from quest.historian import Historian
 from quest.persistence import PersistentHistory, LocalFileSystemBlobStorage
 from .utils import timeout
 from quest.extras.sql import SQLDatabase, SqlBlobStorage
-from quest.extras.dynamodb import DynamoDB, DynamoDBBlobStorage
-from quest.extras.s3 import S3BlobStorage, S3Bucket
+from quest.extras.aws import S3BlobStorage, S3Bucket, DynamoDB, DynamoDBBlobStorage
+
+
+class TestFileSystemStorageContext:
+    def __enter__(self):
+        self.tmp_dir = Path('test_dir')
+        storage = LocalFileSystemBlobStorage(self.tmp_dir)
+        return storage
+
+    def __exit__(self, *args):
+        return True
 
 
 class FileSystemStorageContext:
@@ -61,9 +69,9 @@ class S3StorageContext:
 
 storages = [
     pytest.param(FileSystemStorageContext(), marks=pytest.mark.unit, id="file"),
-    pytest.param(SqlStorageContext(), marks=pytest.mark.unit, id="sql"),
-    pytest.param(DynamoDBStorageContext(), marks=pytest.mark.integration, id="dynamodb"),
-    pytest.param(S3StorageContext(), marks=pytest.mark.integration, id="s3"),
+    # pytest.param(SqlStorageContext(), marks=pytest.mark.unit, id="sql"),
+    # pytest.param(DynamoDBStorageContext(), marks=pytest.mark.integration, id="dynamodb"),
+    # pytest.param(S3StorageContext(), marks=pytest.mark.integration, id="s3"),
 ]
